+++ tests/end_to_end/stack_reference/irix-o2-nocontext-out.c
@@ -1,3 +1,3 @@
 s32 test(s32 arg0, ? arg1) {
-    return (&arg1 - sp) >> 2;
+    return (s32) (&arg1 - sp) >> 2;
 }
